#25 식물 상세 조회 기능 (#26)
* 충돌 해결: App.jsx를 팀원 코드에 맞춤

* Feat: 식물 상세 조회 기능 완료

* 환경변수 관련변경

+++ backend/plantory_be/src/main/java/org/example/plantory_be/entity/PlantDictionary.java
@@ -72,7 +72,7 @@ public class PlantDictionary {
 
     // 테스트용
     @Column(name = "perenual_id")
-    private Integer perenualId;
+    private Long perenualId;
 
     @Column(name = "common_name")
     private String commonName;
